[#2060] Split interpreter into caps & move into separate folder
