+++ msm/src/circuit_design/mod.rs
@@ -1,0 +1,10 @@
+pub mod capabilities;
+pub mod constraints;
+pub mod witness;
+
+// Reexport main types
+pub use crate::circuit_design::{
+    capabilities::{ColAccessCap, ColWriteCap, LookupCap},
+    constraints::ConstraintBuilderEnv,
+    witness::WitnessBuilderEnv,
+};
