+++ msm/src/fec/mod.rs
@@ -8,6 +8,7 @@ pub mod lookups;
 mod tests {
 
     use crate::{
+        circuit_design::{ConstraintBuilderEnv, WitnessBuilderEnv},
         columns::{Column, ColumnIndexer},
         fec::{
             columns::{FECColumn, FEC_N_COLUMNS},
@@ -15,7 +16,6 @@ mod tests {
             lookups::LookupTable,
         },
         prover::prove,
-        serialization::{constraints::ConstraintBuilderEnv, witness::WitnessBuilderEnv},
         verifier::verify,
         witness::Witness,
         BaseSponge, Ff1, Fp, OpeningProof, ScalarSponge, BN254,
